Move swr cache v2 to its own file for now
